rename queuePartnerCommissionCreation, fix tests
